# transportManager/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transport Management's API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/transportAssistant/cityList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "3a2395af58615778ac0db415652a99b7",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/stateList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "3128eaf32146ba25228c0576c17fdb6e",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/goodsList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "051a5bae331f885829408ec791748d7c",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/packagingList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "48b7ba6edc7b49a7dfa29efb2c320523",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/loaderList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "8dfd8a656cee188c48625f7c90262696",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/brandList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "801cd34c5efc32c6f254457857bef578",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/fleetTypeList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "48f9da56c30a44f0a6d9c65bc11365e4",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/transportAssistant/insurerList": {
            "get": {
                "tags": [
                    "Transport Assistant :: Basic Info"
                ],
                "operationId": "aea7b9adde2d2eef3d5f46622beb8d81",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}